refactor(boundary): PR-2 split controller download/runtime + wire-error vocabulary
Remove the agent -> supervisor.controllers import edge so the import-linter
contract can forbid it strictly (4 contracts kept, 0 broken).

- Agent-side downloaders (aleph.vm.agent.vm.downloader: QemuDownloader,
  QemuConfidentialDownloader, ProgramDownloader) compose the host-resource
  surface and raise the contract error vocabulary
  (supervisor_interface.errors.{ResourceDownloadError,FileTooLargeError,
  VmSetupError}). translate.py uses them instead of the controller Resources.
- Controller runtime holders (AlephQemuResources/AlephQemuConfidentialResources/
  AlephFirecrackerResources/AlephProgramResources) lose their now-unused
  download_*/make_writable_volume methods; from_spec stays.
- Shared, neutral foundation modules:
  - aleph.vm.host_volumes (HostVolume, host_volumes_from_message)
  - aleph.vm.program_config (Interface, Volume, ProgramConfiguration,
    ConfigurationPayload*, RunCodePayload, ConfigurationResponse,
    read_input_data, VmInitNotConnectedError)
  - aleph.vm.backup_staging (get_backup_directory, download_volume_by_ref)
  - get_hostname_from_hash moved to aleph.vm.utils
  controllers re-export these for their own importers.
- views/__init__.py + run.py + program_client.py catch/raise the contract
  SupervisorError vocabulary instead of controller/hypervisor exception types;
  operator.py uses the neutral backup-staging utils.
- pyproject: aleph.vm.supervisor.controllers added to the agent contract's
  forbidden_modules; the documented PR-1 residual is gone.
