+++ tests/supervisor/test_supervisor_translate.py
@@ -19,10 +19,11 @@ from aleph_message.models.execution.volume import ParentVolume, VolumePersistenc
 from aleph_message.utils import Mebibytes
 
 from aleph.vm.agent.translate import build_create_vm_spec, build_reservation_request
-from aleph.vm.supervisor.controllers.qemu.cloudinit import get_hostname_from_hash
-from aleph.vm.supervisor.controllers.resources import HostVolume
+from aleph.vm.agent.vm.downloader import QemuDownloader
+from aleph.vm.host_volumes import HostVolume
 from aleph.vm.supervisor_interface.errors import InvalidBackendError
 from aleph.vm.supervisor_interface.types import Backend, DiskRole
+from aleph.vm.utils import get_hostname_from_hash
 
 # ---------------------------------------------------------------------------
 # Helpers
@@ -82,9 +83,7 @@ async def test_build_create_vm_spec_happy_path(monkeypatch: pytest.MonkeyPatch) 
         self.rootfs_path = rootfs
         self.volumes = [vol]
 
-    from aleph.vm.supervisor.controllers.qemu.instance import AlephQemuResources
-
-    monkeypatch.setattr(AlephQemuResources, "download_all", fake_download_all)
+    monkeypatch.setattr(QemuDownloader, "download_all", fake_download_all)
 
     message = _make_qemu_instance_message(
         internet=True,
@@ -146,9 +145,7 @@ async def test_build_create_vm_spec_maps_gpu_requirements(monkeypatch: pytest.Mo
         self.rootfs_path = Path("/data/rootfs.qcow2")
         self.volumes = []
 
-    from aleph.vm.supervisor.controllers.qemu.instance import AlephQemuResources
-
-    monkeypatch.setattr(AlephQemuResources, "download_all", fake_download_all)
+    monkeypatch.setattr(QemuDownloader, "download_all", fake_download_all)
 
     message = _make_qemu_instance_message()
     message = message.model_copy(
@@ -181,9 +178,7 @@ async def test_authorized_keys_none_becomes_empty_list(monkeypatch: pytest.Monke
         self.rootfs_path = Path("/data/rootfs.qcow2")
         self.volumes = []
 
-    from aleph.vm.supervisor.controllers.qemu.instance import AlephQemuResources
-
-    monkeypatch.setattr(AlephQemuResources, "download_all", fake_download_all)
+    monkeypatch.setattr(QemuDownloader, "download_all", fake_download_all)
 
     message = _make_qemu_instance_message(authorized_keys=None)
     spec = await build_create_vm_spec(_VM_HASH, message)
@@ -213,9 +208,7 @@ async def test_non_instance_message_raises(monkeypatch: pytest.MonkeyPatch) -> N
         authorized_keys=None,
     )
 
-    from aleph.vm.supervisor.controllers.qemu.instance import AlephQemuResources
-
-    monkeypatch.setattr(AlephQemuResources, "download_all", should_not_be_called)
+    monkeypatch.setattr(QemuDownloader, "download_all", should_not_be_called)
 
     with pytest.raises(InvalidBackendError, match="InstanceContent"):
         await build_create_vm_spec(_VM_HASH, fake_message)  # type: ignore[arg-type]
@@ -232,9 +225,7 @@ async def test_firecracker_hypervisor_raises(monkeypatch: pytest.MonkeyPatch) ->
         nonlocal download_called
         download_called = True
 
-    from aleph.vm.supervisor.controllers.qemu.instance import AlephQemuResources
-
-    monkeypatch.setattr(AlephQemuResources, "download_all", should_not_be_called)
+    monkeypatch.setattr(QemuDownloader, "download_all", should_not_be_called)
 
     message = _make_qemu_instance_message(hypervisor=HypervisorType.firecracker)
 
@@ -263,9 +254,7 @@ async def test_confidential_instance_populates_tee(monkeypatch: pytest.MonkeyPat
     async def fake_get_existing_file(ref: Any) -> Path:
         return firmware_path
 
-    from aleph.vm.supervisor.controllers.qemu.instance import AlephQemuResources
-
-    monkeypatch.setattr(AlephQemuResources, "download_all", fake_download_all)
+    monkeypatch.setattr(QemuDownloader, "download_all", fake_download_all)
     monkeypatch.setattr("aleph.vm.agent.translate.get_existing_file", fake_get_existing_file)
 
     # The message requests a non-default policy: the spec must carry that
@@ -300,9 +289,7 @@ async def test_hypervisor_none_defaults_to_qemu(monkeypatch: pytest.MonkeyPatch)
         self.rootfs_path = rootfs
         self.volumes = []
 
-    from aleph.vm.supervisor.controllers.qemu.instance import AlephQemuResources
-
-    monkeypatch.setattr(AlephQemuResources, "download_all", fake_download_all)
+    monkeypatch.setattr(QemuDownloader, "download_all", fake_download_all)
     monkeypatch.setattr("aleph.vm.conf.settings.INSTANCE_DEFAULT_HYPERVISOR", HypervisorType.qemu)
 
     message = _make_qemu_instance_message(hypervisor=None)
@@ -323,9 +310,7 @@ async def test_hypervisor_none_defaults_to_firecracker_raises(monkeypatch: pytes
         nonlocal download_called
         download_called = True
 
-    from aleph.vm.supervisor.controllers.qemu.instance import AlephQemuResources
-
-    monkeypatch.setattr(AlephQemuResources, "download_all", should_not_be_called)
+    monkeypatch.setattr(QemuDownloader, "download_all", should_not_be_called)
     monkeypatch.setattr("aleph.vm.conf.settings.INSTANCE_DEFAULT_HYPERVISOR", HypervisorType.firecracker)
 
     message = _make_qemu_instance_message(hypervisor=None)
